requirements.txt
annotated-types==0.5.0
anyio==4.0.0
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
copier==8.3.0
cryptography==41.0.4
decorator==5.1.1
dunamai==1.19.0
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
flet-core==0.10.3
flet-runtime==0.10.3
funcy==2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
paramiko==3.3.1
pathspec==0.11.2
plumbum==1.8.2
prompt-toolkit==3.0.36
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyNaCl==1.5.0
pypng==0.20220715.0
PySocks==1.7.1
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
repath==0.9.0
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.5
watchdog==3.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0